:truck: installable wheel package
